requirements.txt
elasticsearch>=7.13.0
python-dateutil>=2.8.1
pyyaml>=5.4.1